core/analysis/advanced_detection.py
import cv2
import torch
from sort import Sort

class AdvancedAnalyzer:
    def __init__(self):
        self.model = torch.hub.load('ultralytics/yolov5', 'yolov5x6')
        self.tracker = Sort()
        
    def analyze_video(self, path):
        cap = cv2.VideoCapture(path)
        while cap.isOpened():
            ret, frame = cap.read()
            results = self.model(frame)
            tracked_objects = self.tracker.update(results.xyxy[0].cpu().numpy())
            yield self.process_frame(frame, tracked_objects)

    def process_frame(self, frame, objects):
        # Process the frame and objects
        pass