ui.py
import tkinter as tk
from tkinter import ttk
from PIL import Image, ImageTk
import sv_ttk

class FootballAnalysisApp(tk.Tk):
    def __init__(self):
        super().__init__()
        self.title("Football Analysis Pro")
        self.geometry("1366x768")
        
        # Set dark theme
        sv_ttk.set_theme("dark")
        
        # Configure style
        self.style = ttk.Style()
        self.style.configure("TButton", padding=6, font=("Roboto", 10))
        self.style.map("Accent.TButton",
            foreground=[('active', 'white'), ('!active', 'white')],
            background=[('active', '#1E88E5'), ('!active', '#1565C0')]
        )
        
        # Main container
        self.container = ttk.Frame(self)
        self.container.pack(fill=tk.BOTH, expand=True)
        
        # Initialize components
        self.create_dashboard()
        self.create_video_panel()
        self.create_stats_hub()
        
    def create_dashboard(self):
        # Dashboard frame with summary cards
        self.dashboard_frame = ttk.Frame(self.container)
        
        # Summary cards
        summary_metrics = ["xG", "Possession", "Shots"]
        for i, metric in enumerate(summary_metrics):
            card = ttk.Frame(self.dashboard_frame, style="Card.TFrame")
            ttk.Label(card, text=metric, style="CardTitle.TLabel").pack()
            ttk.Label(card, text="0.0", style="CardValue.TLabel").pack()
            card.grid(row=0, column=i, padx=10, pady=10, sticky="nsew")
        
        # Recent analyses
        recent_frame = ttk.Frame(self.dashboard_frame)
        ttk.Label(recent_frame, text="Recent Analyses").pack()
        # Add thumbnail previews
        
        self.dashboard_frame.pack(fill=tk.BOTH, expand=True)

    def create_video_panel(self):
        # Split-screen video analysis
        self.video_panel = ttk.PanedWindow(self.container, orient=tk.HORIZONTAL)
        
        # Video player
        self.video_frame = ttk.Frame(self.video_panel, width=800)
        self.video_canvas = tk.Canvas(self.video_frame)
        self.video_canvas.pack(fill=tk.BOTH, expand=True)
        
        # Analysis panel
        self.analysis_frame = ttk.Frame(self.video_panel, width=400)
        self.insights_text = tk.Text(self.analysis_frame, wrap=tk.WORD)
        self.insights_text.pack(fill=tk.BOTH, expand=True)
        
        self.video_panel.add(self.video_frame)
        self.video_panel.add(self.analysis_frame)

    def create_stats_hub(self):
        # Interactive statistics
        self.stats_notebook = ttk.Notebook(self.container)
        
        # Main stats tab
        self.stats_frame = ttk.Frame(self.stats_notebook)
        self.filter_frame = ttk.Frame(self.stats_frame)
        ttk.Combobox(self.filter_frame, values=["Defensive", "Offensive"]).pack()
        
        # Visualization canvas
        self.stats_canvas = tk.Canvas(self.stats_frame)
        self.stats_canvas.pack(fill=tk.BOTH, expand=True)
        
        self.stats_notebook.add(self.stats_frame, text="Performance Analytics")
        self.stats_notebook.pack(fill=tk.BOTH, expand=True)

    # Add remaining functionality...

if __name__ == "__main__":
    app = FootballAnalysisApp()
    app.mainloop()