requirements.txt
customtkinter==5.2.1
sv-ttk==2.5.0
Pillow==10.3.0
opencv-python==4.9.0
torch==2.2.1
torchvision==0.17.1
pandas==2.2.1
fpdf2==2.8.2
requests==2.31.0
ultralytics
sort
moviepy==1.0.3
imageio==2.37.0
imageio-ffmpeg==0.4.9
ffmpeg-python
flask
flask-restful
pyinstaller
scikit-learn
matplotlib
numpy